Created method to find the major (and technically minor) coordinate of the cube from a given size
if a user wants a cube of size 3 the top corner would have coordinate (1.5,1.5,1.5) (half the size of the cube). if you want to apply an offset to the cube to have it on the positive side of the axis you could just add this value as an offset. an example of this is in ThreeByThreePositive()