Update generic manipualtion to use rotors
The rotation demonstraition no longer works now each shader users
rotor rotation - so it was updated to use rotors. Interesting, but
debatably good consequence of this - if you uncheck a rotation, it
does not reset the rotation, it maintains it. could be good, could
be frustarting. These scenes are not really used anyway. Just for
maintenance.